# api/openapi.yaml
openapi: 3.1.0
info:
  title: Trident REST API
  description: |
    Real-time indexing and event querying for Soroban contract events.

    ## Request validation

    Every endpoint validates its parameters before doing any work and answers a
    malformed request with `400` and the `ErrorResponse` envelope, where
    `error.code` is `INVALID_ARGUMENT` and `error.message` names the field that
    failed:

    ```json
    {"error": {"code": "INVALID_ARGUMENT", "message": "limit must be an integer between 1 and 200"}}
    ```

    Rules shared across endpoints:

    - `contractId` — Stellar contract strkey: `C` followed by 55 uppercase
      base32 characters.
    - path ids — canonical UUID v4.
    - ledger ranges — non-negative integers; the upper bound must be greater
      than or equal to the lower bound.
    - `limit` — an integer within the range documented for that endpoint.
    - `cursor` — an opaque cursor from a previous response; anything else is
      rejected rather than silently ignored.
    - `network` — `testnet` or `mainnet`.
    - Unknown query parameters are rejected, so a typo such as `?limitt=5`
      fails loudly instead of silently changing the page size.
  version: 1.0.0
  contact:
    name: Trident Team
  license:
    name: Apache 2.0
servers:
  - url: https://api.trident.telocel.com
    description: Production
  - url: http://localhost:3000
    description: Development
security:
  - ApiKeyAuth: []
paths:
...
  /v1/api-keys:
    post:
      summary: Create an API key
      description: Generate a new API key for authentication
      operationId: createApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                label:
                  type: string
                  description: Human-readable label for this key
                network:
                  type: string
                  enum: [testnet, mainnet]
                  default: mainnet
                  description: Network this key is associated with
                rate_limit_tier:
                  type: string
                  enum: [free, standard, premium, internal]
                  default: standard
                  description: Rate limit tier for this key
      responses:
        "201":
          description: API key created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  key:
                    type: string
                    description: The raw API key (shown only once)
                  key_prefix:
                    type: string
                    description: First 16 characters of key for reference
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"

    get:
      summary: List API keys
      description: List all API keys (admin only), using keyset pagination consistent with GET /v1/events
      operationId: listApiKeys
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
          description: Maximum number of API keys to return
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque pagination cursor from previous response's next_cursor (for next page)
      responses:
        "200":
          description: List of API keys with pagination metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        key_prefix:
                          type: string
                        label:
                          type: string
                        network:
                          type: string
                        rate_limit_tier:
                          type: string
                        last_used_at:
                          type: string
                          format: date-time
                        request_count:
                          type: integer
                        created_at:
                          type: string
                          format: date-time
                  has_more:
                    type: boolean
                    description: Whether more pages are available
                  next_cursor:
                    type: string
                    nullable: true
                    description: Opaque cursor to pass as the cursor parameter for the next page
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"
  /v1/api-keys/{id}:
    delete:
      summary: Delete an API key
      description: Revoke and delete an API key (admin only)
      operationId: deleteApiKey
      tags:
        - API Keys
      security:
        - AdminKey: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: API key ID
      responses:
        "204":
          description: API key deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: API key not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/TooManyRequestsIPOnly"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code (e.g., INVALID_ARGUMENT, INTERNAL, UNAVAILABLE)
            message:
              type: string
              description: Human-readable error message
            request_id:
              type: string
              description: Request ID for debugging
  headers:
...
    Retry-After:
      description: >-
        Seconds to wait before retrying. Present on 429 (rate limit
        exceeded, per-API-key or per-IP) and on 503 responses caused by the
        global concurrency cap shedding load. Not present on a 503 caused by
        an unavailable dependency (database/Redis/gRPC backend) — check the
        error envelope's `error.code` to distinguish the two.
      schema:
        type: integer
        minimum: 1
      example: 1
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    TooManyRequestsIPOnly:
      description: >-
        Per-IP rate limit exceeded (error.code RATE_LIMITED). Applies to
        endpoints not covered by per-API-key limiting (public endpoints, or
        admin endpoints authenticated via ADMIN_API_KEY rather than
        X-API-Key) — only Retry-After is set, no X-RateLimit-* headers.
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    AdminKey:
      type: apiKey
      in: header
      name: ADMIN_API_KEY
      description: Admin key for admin endpoints (environment variable)
...